% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-cleanup.R
\name{ce_cleanup}
\alias{ce_cleanup}
\title{Clean up directory used to store CE PUMD}
\usage{
ce_cleanup()
}
\arguments{
\item{ce_dir}{The directory in which CE PUMD data and metadata are stored. If
\code{NULL} (the default) a \code{ce_cleanup} will look for a directory
called "ce_data" to delete it.}
}
\description{
A function to delete the directory containing CE PUMD. It
will first check whether the provided directory exists and it will avoid
deleting the main temporary directory.
}
