% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-mean.R
\name{ce_mean}
\alias{ce_mean}
\title{Calculate a CE weighted, estimated mean}
\usage{
ce_mean(ce_data)
}
\arguments{
\item{ce_data}{A data frame containing at least a finlwt21 column,
44 replicate weight columns (wtrep01-44), and a cost column. All of these
columns must be numeric.}
}
\value{
A 1-row dataframe containing the following columns:
\itemize{
  \item agg_exp - The estimated aggregate expenditure
  \item mean_exp - The estimated mean expenditure
  \item se - The estimated standard error of the estimated mean expenditure
  \item cv - The coefficient of variation of the estimated mean expenditure
}
}
\description{
Calculate a CE weighted, estimated mean
}
\examples{
# Calculate the mean pet expenditure using only Diary expenditures
ce_mean(ce_diary_pets17)
}
\seealso{
\code{\link{ce_median}}
}
