% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-median.R
\name{ce_median}
\alias{ce_median}
\title{Calculate a CE weighted, estimated median}
\usage{
ce_median(ce_data)
}
\arguments{
\item{ce_data}{A data frame containing at least a finlwt21 column and a cost
column. Both columns must be numeric.}
}
\value{
A weighted, estimated median (numeric)
}
\description{
Calculate a CE weighted, estimated median
}
\examples{
# Calculate the mean pet expenditure
ce_median(ce_diary_pets17)
}
\seealso{
\code{\link{ce_mean}}
}
