% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-pumd-years.R
\name{ce_pumd_years}
\alias{ce_pumd_years}
\title{Get a vector of years for which PUMD are available}
\usage{
ce_pumd_years()
}
\value{
A numeric vector (years)
}
\description{
Scrapes the CE PUMD website for the table of CSV files and
pulls the column of that table containing the years.
}
