% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-quantiles.R
\name{ce_quantiles}
\alias{ce_quantiles}
\title{Calculate a CE weighted quantiles}
\usage{
ce_quantiles(ce_data, probs = 0.5)
}
\arguments{
\item{ce_data}{A data frame containing at least a finlwt21 column and a cost
column. Both columns must be numeric.}

\item{probs}{A numeric vector of probabilities between 0 and 1 for which to
compute quantiles. Default is 0.5 (median).}
}
\value{
A two-column data frame in which the first column contains the
probabilities for which quantiles were calculated and their corresponding
quantiles in the second column.
}
\description{
Calculate a CE weighted quantiles
}
\examples{
# Calculate the median pet expenditure
ce_quantiles(ce_diary_pets17, c(0.25, 0.5, 0.75))
}
\seealso{
\code{\link{ce_mean}}
}
