% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-stub.R
\name{ce_stub}
\alias{ce_stub}
\title{Generate a CE stub file as data frame}
\usage{
ce_stub(year, stub_type)
}
\arguments{
\item{year}{The year corresponding to the stub file of interest.}

\item{stub_type}{The type of stub file, i.e., interview, diary, or
integrated. Accepted as a string or symbol (quotes or no quotes).}
}
\value{
A data frame containing the following columns:
* level - hierarchical level of the expenditure category
* title - the title of the expenditure category
* ucc - the Universal Classification Code (UCC) for the expenditure category
* factor - the factor by which to multiply the expenditure in the calculation
  of estimated means / medians
}
\description{
Reads in a CE stub file for the given year and stub type
as data frame.
}
\details{
The CE did not have publicly available stub files between 1996 and 2001.
Instead, they had label files and crosswalk files. The label files had line
line numbers indicating the order of the categories in the title column and
the crosswalk file contained line numbers and UCC's. The titles in the,
labels files, however, were the titles for high level category groupings.
The titles for the lower level UCC's were published in the documentation for
the given year of data. I have copied those sections of the documentation
into data files contained in this package, e.g., "interview_uccs_1996."
This function pulls the crosswalk and labels from the CE website, merges
them, then merges in the UCC level titles from the data files.
}
\examples{
# 'stub_type' can be entered as a string
ce_stub(2016, "integrated")

# 'stub_type' can also be entered as a symbol
ce_stub(2016, integrated)
}
