% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-uccs.R
\name{ce_uccs}
\alias{ce_uccs}
\title{Find UCCs for expenditure categories}
\usage{
ce_uccs(hg, expenditure, uccs_only = TRUE)
}
\arguments{
\item{hg}{A data frame that has, at least, the title, level, and ucc
columns of a CE HG file.}

\item{expenditure}{A string that is an expenditure category contained in a
CE HG file (exact match required).}

\item{uccs_only}{A logical indicating whether to return only the expenditure
category's component ucc's. If TRUE (default), a vector of UCC's will be
returned. If FALSE, a dataframe will be returned containing the section of
the HG file containing the expenditure category and its component sub-
categories}
}
\value{
A vector of Universal Classification Codes (UCC's) corresponding to
the lowest hierarchical level for that category.
}
\description{
Find UCCs for expenditure categories
}
\examples{
# First generate an HG file
my_hg <- ce_hg(2017, interview)

# Store a vector of UCC's in the "Pets" category
pet_uccs <- ce_uccs(my_hg, "Pets")
pet_uccs
# [1] "610320" "620410" "620420"
}
