% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepumd-package.R
\docType{package}
\name{cepumd-package}
\alias{cepumd}
\alias{cepumd-package}
\title{cepumd: Calculate Consumer Expenditure Survey (CE) Annual Estimates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions and data files to help CE Public-Use Microdata (PUMD) users calculate annual estimated expenditure means, standard errors, and quantiles according to the methods used by the CE with PUMD. For more information on the CE please visit \url{https://www.bls.gov/cex}. For further reading on CE estimate calculations please see the CE Calculation section of the U.S. Bureau of Labor Statistics (BLS) Handbook of Methods at \url{https://www.bls.gov/opub/hom/cex/calculation.htm}. For further information about CE PUMD please visit \url{https://www.bls.gov/cex/pumd.htm}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arcenis-r.github.io/cepumd/}
  \item \url{https://github.com/arcenis-r/cepumd}
  \item Report bugs at \url{https://github.com/arcenis-r/cepumd/issues}
}

}
\author{
\strong{Maintainer}: Arcenis Rojas \email{arcenis.rojas@gmail.com} [copyright holder]

}
\keyword{internal}
