% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-survey-files.R
\name{get_survey_files}
\alias{get_survey_files}
\title{Generate tables of the necessary survey data files}
\usage{
get_survey_files(year, survey, file_yrs, qtrs, ce_dir, zp_file)
}
\arguments{
\item{year}{A year between 1996 and the last year of available CE PUMD.}

\item{survey}{One of either interview, diary, or integrated as a character or
symbol.}

\item{file_yrs}{The substrings of years for which to pull data, i.e., for
some years files have to be pulled from across different files.}

\item{qtrs}{The quarters to be included in the analysis for a given year.}

\item{ce_dir}{The directory in which CE PUMD data and metadata are stored. If
\code{NULL} (the default) a directory called "ce-data" will be created in the
temporary directory of the session.}

\item{zp_file}{Character indicating the zip file containing the CE PUMD for a
given year}
}
\description{
Generate tables of the necessary survey data files
}
\details{
This is a hidden file called only by exported package functions.
}
