% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-stub-96-01.R
\name{make.stub.96.01}
\alias{make.stub.96.01}
\title{Generate a stub from a year between 1996 and 2001}
\usage{
make.stub.96.01(year, stub_type_name)
}
\arguments{
\item{year}{Year corresponding to the data}

\item{stub_type_name}{Stub file type}
}
\description{
Generate a stub from a year between 1996 and 2001
}
\details{
This is a hidden file called only by exported package functions.
}
