% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-expd.R
\name{read.expd}
\alias{read.expd}
\title{Read in and modify EXPD files}
\usage{
read.expd(fp, zp, year, uccs, integrate_data, hg, ce_dir)
}
\arguments{
\item{fp}{File to extract from zip file}

\item{zp}{Zip file path}

\item{year}{Year}

\item{uccs}{Vector of UCC's to filter for}

\item{integrate_data}{Whether to prepare data for integrated estimates}

\item{hg}{Hierarchical grouping data}
}
\description{
Read in and modify EXPD files
}
\details{
This is a hidden file called only by exported package functions.
}
