% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fmli.R
\name{read.fmli}
\alias{read.fmli}
\title{Read in and modify FMLI files}
\usage{
read.fmli(fp, zp, year, grp_var_names)
}
\arguments{
\item{fp}{File to extract from zip file}

\item{zp}{Zip file path}

\item{year}{Year}

\item{grp_var_names}{Variables to keep (intended for grouping)}
}
\description{
Read in and modify FMLI files
}
\details{
This is a hidden file called only by exported package functions.
}
