% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-ce-variables.R
\name{recode_ce_variables}
\alias{recode_ce_variables}
\title{Recode variables in interview and diary data}
\usage{
recode_ce_variables(srvy_data, code_file, srvy)
}
\arguments{
\item{srvy_data}{A data frame containing either Interview or Diary data that
has been prepped}

\item{code_file}{A dataframe containing variable names, codes,
code descriptions, and other required columns for recoding variables}

\item{srvy}{The survey instrument to be recoded (this is for filtering
the codebook)}
}
\description{
Recode variables in interview and diary data
}
\details{
This is a hidden file called only by exported package functions.
}
