% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-ce-dict.R
\name{store_ce_dict}
\alias{store_ce_dict}
\title{Download CE Public-Use Microdata Dictionary}
\usage{
store_ce_dict(ce_dir = NULL, dict_path = NULL)
}
\arguments{
\item{ce_dir}{The directory in which CE PUMD data and metadata are stored. If
\code{NULL} (the default) a directory called "ce-data" will be created in the
temporary directory of the session.}

\item{dict_path}{A valid file path entered as a string to which to store a
the CE PUMD Excel dictionary. If \code{NULL} (the default) a file called
"ce-stubs.zip" will be created in \code{ce_dir}.}
}
\description{
A wrapper for \code{\link[utils]{download.file}} that downloads
the CE PUMD dictionary as and Excel file to the path indicated in the
\code{hg_zip_path} argument.
}
\details{
The download comes from
\url{https://www.bls.gov/cex/pumd/ce_pumd_interview_diary_dictionary.xlsx}
to the path designated in \code{hg_zip_path} using \code{method = "wb"}
}
\examples{
# This will download the dictionary in the current working directory
\dontrun{
store_ce_dict("ce-dict.xlsx")

# Check the file
file.info("ce-dict.xlsx")
}
}
