% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-ce-hg.R
\name{store_ce_hg}
\alias{store_ce_hg}
\title{Download CE hierarchical grouping files}
\usage{
store_ce_hg(ce_dir = NULL, hg_zip_path = NULL)
}
\arguments{
\item{ce_dir}{The directory in which CE PUMD data and metadata are stored. If
\code{NULL} (the default) a directory called "ce-data" will be created in the
temporary directory of the session.}

\item{hg_zip_path}{A valid file path entered as a character to which to store
a zip file containing the hierarchical grouping files. If \code{NULL} (the
default) a file called "ce-stubs.zip" will be created in \code{ce_dir}.}
}
\description{
A wrapper for \code{\link[utils]{download.file}} that downloads
a zip file containing all CE hieararchical grouping files (also known as
'stub' files) to the path indicated in the \code{hg_zip_path} argument.
}
\details{
The download comes from
\url{https://www.bls.gov/cex/pumd/stubs.zip} to the path designated in
\code{hg_zip_path} using \code{method = "wb"}
}
\examples{
# This will download the zip file in the current working directory
\dontrun{
store_ce_hg("ce-hg-files.zip")

# Check the file
file.info("ce-hg-files.zip")

head(unzip("ce-hg-files.zip", list = TRUE))
}
}
