% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-download.R
\name{ce_download}
\alias{ce_download}
\title{Download CE zip files}
\usage{
ce_download(year, survey, zp)
}
\arguments{
\item{year}{The year corresponding to the data to download.}

\item{survey}{One of either "interview" or "diary" as a string or symbol.}

\item{zp}{A string indicating the path where you'd like to save the zip file
(must end in ".zip").}
}
\value{
Stores a zip file containing CE data in the designated location.
}
\description{
Downloads zip files directly from the CE PUMD page on the
    U.S. Bureau of Labor Statistics website to the path (zp) designated by
    the user.
}
\examples{
# 'survey' can be entered as a string
ce_download(2017, "interview", "./interview17.zip")

# 'survey' can also be entered as a symbol
ce_download(2017, interview, "./interview17.zip")
}
