% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-mean.R
\name{ce_mean}
\alias{ce_mean}
\title{Calculate a CE weighted, estimated mean}
\usage{
ce_mean(ce_data)
}
\arguments{
\item{ce_data}{A data frame containing a finlwt21 column, 44 replicate weight
columns (wtrep01-44), and a cost column. All data must be numeric.}
}
\value{
A 1-row dataframe containing the following columns:
agg_exp - The estimated aggregate expenditure
mean_exp - The estimated mean expenditure
se - The estimated standard error of the estimated mean expenditure
cv - The coefficient of variation of the estimated mean expenditure
}
\description{
Calculate a CE weighted, estimated mean
}
\examples{
# The following workflow will prepare a dataset for calculating integrated
# pet expenditures for 2017 and calculate the estimated mean expenditure and
# related summary statistics.

# Calculate the mean pet expenditure
ce_mean(ce_pets17)
}
