% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce-uccs.R
\name{ce_uccs}
\alias{ce_uccs}
\title{Find the UCC's that correspond to expenditure categories}
\usage{
ce_uccs(stub, expenditure)
}
\arguments{
\item{stub}{A data frame that has, at least, the title, level, and ucc
columns of a CE stub file.}

\item{expenditure}{An expenditure category exactly matching the way it is
written in a CE stub file (as a string).}
}
\value{
A vector of Universal Classification Codes (UCC's) corresponding to
the lowest hierarchical level for that category.
}
\description{
Find the UCC's that correspond to expenditure categories
}
\examples{
# First generate a stub file
mystub <- ce_stub(2017, interview)

# Store a vector of UCC's in the "Pets" category
pet_uccs <- ce_uccs(mystub, "Pets")
pet_uccs
# [1] "610320" "620410" "620420"
}
