% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepumd-package.R
\docType{package}
\name{cepumd-package}
\alias{cepumd}
\alias{cepumd-package}
\title{cepumd: Calculate Consumer Expenditure Survey (CE) Annual Estimates}
\description{
Provides functions and data files to help CE PUMD users calculate annual estimated expenditure means, standard errors, and quantiles according to the methods used by the CE using Public-Use Microdata (PUMD). For more information on the CE please visit \url{https://www.bls.gov/cex}. For further reading on CE estimate calculations please see the CE Calculation section of the BLS Handbook of Methods at \url{https://www.bls.gov/opub/hom/cex/calculation.htm}. For further informaiton about CE PUMD please visit \url{https://www.bls.gov/cex/pumd.htm}.
}
\keyword{internal}
