% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fmld.R
\name{read.fmld}
\alias{read.fmld}
\title{Read in and modify FMLD files}
\usage{
read.fmld(fp, zp, grp_var_names, ce_dir)
}
\arguments{
\item{fp}{File to extract from zip file}

\item{zp}{Zip file path}

\item{grp_var_names}{Variables to keep (intended for grouping)}
}
\description{
Read in and modify FMLD files
}
\details{
This is a hidden file called only by exported package functions.
}
