% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fmli.R
\name{read.fmli}
\alias{read.fmli}
\title{Read in and modify FMLI files}
\usage{
read.fmli(fp, zp, year, ce_dir, grp_var_names)
}
\arguments{
\item{fp}{File to extract from zip file}

\item{zp}{Zip file path within ce_dir}

\item{year}{Year}

\item{ce_dir}{The directory in which CE PUMD data and metadata are stored. If
\code{NULL} (the default) a directory called "ce-data" will be created in the
temporary directory of the session.}

\item{grp_var_names}{Variables to keep (intended for grouping)}
}
\description{
Read in and modify FMLI files
}
\details{
This is a hidden file called only by exported package functions.
}
