% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fmli.R
\name{read.fmli}
\alias{read.fmli}
\title{Read in and modify FMLI files}
\usage{
read.fmli(fp, zp, year, ...)
}
\arguments{
\item{fp}{File to extract from zip file}

\item{zp}{Zip file path within ce_dir}

\item{year}{Year}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional variables to keep
(intended for grouping)}
}
\description{
Read in and modify FMLI files
}
\details{
This is a hidden file called only by exported package functions.
}
